src/test/java/ru/yandex/practicum/filmorate/testFilm/FilmDbStorageTest.java
package ru.yandex.practicum.filmorate.testFilm;


import lombok.RequiredArgsConstructor;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.jdbc.core.JdbcTemplate;
import ru.yandex.practicum.filmorate.model.Film;
import ru.yandex.practicum.filmorate.model.Rating;
import ru.yandex.practicum.filmorate.storage.film.FilmDbStorage;

import java.time.LocalDate;
import java.util.LinkedHashSet;

import static org.junit.jupiter.api.Assertions.assertFalse;
import static org.junit.jupiter.api.Assertions.assertTrue;


@SpringBootTest
@AutoConfigureTestDatabase
@RequiredArgsConstructor(onConstructor_ = @Autowired)
public class FilmDbStorageTest {

    Film testFilm1;
    @Autowired
    private FilmDbStorage filmDbStorage;

    @Autowired
    private JdbcTemplate jdbcTemplate;

    @BeforeEach
    public void beforeEach() {
        jdbcTemplate.execute("delete from films");

        testFilm1 = new Film(1, "The Matrix", "la-la-la", LocalDate.of(1999, 01, 01),
                200, new Rating(1, "G"), new LinkedHashSet<>());
    }

    @Test
    public void testIsContainFilm() {
        long nonExistingFilmId = 999;

        filmDbStorage.addFilm(testFilm1);

        boolean existingFilmResult = filmDbStorage.isContainFilm(testFilm1.getId());
        boolean nonExistingFilmResult = filmDbStorage.isContainFilm(nonExistingFilmId);

        assertTrue(existingFilmResult);
        assertFalse(nonExistingFilmResult);
    }
}